把框架部分全部移到util里，降低用户理解成本
